Fix for WFMP-258, Refactor Package Mojo to allow for better re-use
